split up the screen_buffer module
